+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- resource-definition clone: new `--volume-size` option (server REST API 1.29.1)
+
 ## [1.29.0] - 2026-08-13
 
 ### Added